--- nexthw/Components/Layout/Layout.props.ts
@@ -1,5 +1,0 @@
-import { ReactNode } from "react";
-
-export interface LayoutProps{
-   children: ReactNode
-}
