+++ nexthw/app/layout.module.css
@@ -4,3 +4,23 @@
     align-items: center;
     padding: 1em;
 }
+
+.wrapper{
+    display: grid;
+    grid-template-columns: auto 1fr 1fr 1fr auto;
+    grid-template-areas: ". wrapper wrapper wrapper .";
+    gap: 20px;
+}
+
+@media (max-width: 800px){
+    .wrapper{
+        grid-template-columns: 1fr 1fr;
+        grid-template-areas: "wrapper wrapper";
+    }
+}
+@media (max-width: 450px){
+    .wrapper{
+        grid-template-columns: 1fr;
+        grid-template-areas: "wrapper";
+    }
+}
